+++ CHANGELOG.rst
@@ -23,6 +23,8 @@ Added
     - Project member notifications from zone moving (#1203, #1232)
     - ``LandingZone.user_message`` field (#1203)
     - ``finished`` parameter for ``LandingZoneListAPIView`` to access finished zones (#1234)
+    - ``LandingZone.is_locked()`` helper (#321)
+    - Zone locked status in UI and ``LandingZoneRetrieveAPIView`` (#321)
 - **Samplesheets**
     - Simple link support for string cell rendering (#1001)
     - ``generic_raw`` assay plugin (#1128)